backend/requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
chromadb
PyPDF2
requests
python-multipart
google-api-python-client
youtube-transcript-api
